+++ src/clj/rems/css/style_utils.clj
@@ -2,13 +2,32 @@
   "The namespace contains the helpful style utils that are meant to be kept separate
    the main styles file to keep it cleaner."
   (:require [clojure.string :as str]
-            [garden.stylesheet :as stylesheet]
             [rems.config :refer [env]]))
 
 
-;; Customazable theme related functions
+;; Customizable theme related functions
 
-(defn get-theme-attribute
+(def ^:private ignore-theme-var-error?
+  ;; We don't want to include all these localized versions in
+  ;; config-defaults.edn. Luckily there are only a few so we can just
+  ;; have a whitelist.
+  #{:logo-name-fi
+    :logo-name-fi-sm
+    :navbar-logo-name-fi
+    :logo-name-sv
+    :logo-name-sv-sm
+    :navbar-logo-name-sv})
+
+(defn- theme-getx-impl
+  "Helper for making sure we document all our theme variables."
+  [attr]
+  (if-let [[_ v] (find (:theme env) attr)] ; find instead of get: nil values are ok, missing values are bad
+    v
+    (when (:dev env)
+      (when-not (ignore-theme-var-error? attr)
+        (assert false (str "Theme attribute " attr " used but not documented in config-defaults.edn!"))))))
+
+(defn theme-getx
   "Fetch the attribute value from the current theme with fallbacks.
 
   Keywords denote attribute lookups while strings are interpreted as fallback constant value."
@@ -16,7 +35,7 @@
   (when (seq attr-names)
     (let [attr-name (first attr-names)
           attr-value (if (keyword? attr-name)
-                       (get (:theme env) attr-name)
+                       (theme-getx-impl attr-name)
                        attr-name)]
       (or attr-value (recur (rest attr-names))))))
 
@@ -24,15 +43,14 @@
   (when path
     (let [url (if (str/starts-with? path "http")
                 path
-                (str (get-theme-attribute :img-path "../../img/") path))]
+                (str (theme-getx :img-path) path))]
       (str "url(\"" url "\")"))))
 
 (defn get-logo-image [lang]
-  (resolve-image (get-theme-attribute (keyword (str "logo-name-" (name lang))) :logo-name)))
+  (resolve-image (theme-getx (keyword (str "logo-name-" (name lang))) :logo-name)))
 
 (defn get-logo-name-sm [lang]
-  (resolve-image (get-theme-attribute (keyword (str "logo-name-" (name lang) "-sm")) :logo-name-sm)))
+  (resolve-image (theme-getx (keyword (str "logo-name-" (name lang) "-sm")) :logo-name-sm)))
 
 (defn get-navbar-logo [lang]
-  (resolve-image (get-theme-attribute (keyword (str "navbar-logo-name-" (name lang))) :navbar-logo-name)))
-
+  (resolve-image (theme-getx (keyword (str "navbar-logo-name-" (name lang))) :navbar-logo-name)))